fix: agent functions into tools
